feat: add example app themes and showcases screens
